Rename kernel_modules to system_modules
